+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 ### Removed
 
 * sdk: remove `nwc` dep ([Yuki Kishimoto])
+* nwc: remove `thiserror` and unnecessary `Error::Zapper` variant ([Yuki Kishimoto])
 * ffi: drop support for `i686-linux-android` target ([Yuki Kishimoto])
 * ffi: remove `MockRelay` ([Yuki Kishimoto])
 * ffi: remove `RawEvent` and `RawEventRecord` ([Yuki Kishimoto])